Добавляет классÑ пророÐдителя Component
